Refactored code, wrote TimerSetItems object
Wrote the TimerSetItems object to be responsible for setting items repeatedly i.e. magic, health which are set every 50 ms. Items can now be added with a single line function call. The LBA1 code was then re-written to use this new module also as well as implementing inf health/magic for LBA2